Move solver/bounded.rs to bounded/set.rs
This aligns with the changes that I made to sat-checker and bdd-checker.
