+++ src/bounded/mod.rs
@@ -1,0 +1,6 @@
+// Copyright 2022-2023 VMware, Inc.
+// SPDX-License-Identifier: BSD-2-Clause
+
+//! A collection of bounded model checkers.
+
+pub mod set;
